Fix user tracking for users outside channels
Previously, if an unknown user sent a PRIVMSG directly to the
connection, it would become tracked. When a user with the same name then
joined a tracked channel, they would be treated as if they were the
first user. That is not necessarily the case.

Also, channels the connection is no longer in would stay alive in its
list of channels, even though nothing is actually known about them. This
would also keep user objects alive, since they were still seen as being
in that channel.

This commit fixes the issues and adds some tests around user tracking.